requirements.txt
pytest==3.5.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
requests-html==0.9.0
requests-toolbelt==0.8.0
ruamel.yaml==0.15.80
selenium==3.0.1
six==1.11.0
sqlmap==1.2.7
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.5
yagmail==0.10.212